implemented JPA
