+++ src/main/java/co/inventorsoft/mailsecurity/models/Email.java
@@ -5,24 +5,22 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.format.annotation.DateTimeFormat;
 
+import javax.persistence.*;
 import java.time.LocalDateTime;
 
 @NoArgsConstructor
 @AllArgsConstructor
+@Entity
+@Table(name = "emails")
 @Data
 public class Email {
-    private int id;
+    @Id
+    @GeneratedValue
+    private Long id;
     private String recipient;
     private String subject;
     private String body;
 
     @DateTimeFormat(pattern = "dd.MM.yyyy HH:mm")
     private LocalDateTime date;
-
-    public Email(String recipient, String subject, String body, LocalDateTime date) {
-        this.recipient = recipient;
-        this.subject = subject;
-        this.body = body;
-        this.date = date;
-    }
 }
